Call the GetCutOffDatesLambda logic from the main lambda
